+++ docs/demos/FormGeneratorDemo2.tsx
@@ -1,6 +1,3 @@
-import React from 'react';
-// import { Button, Form, Input, SelectProps, Space, TimeRangePickerProps } from 'antd';
-// import { FormGenerator, FormItemConfig, RenderType } from 'art-antd-react';
 //  安裝了 art-antd-react 以后 antd 直接导出的 类型或组件 可以直接从 art-antd-react 中导出
 import {
   Button,