+++ docs/demos/FormModalDemo1.tsx
@@ -1,13 +1,13 @@
-import { Button, Form, message } from 'antd';
 import {
+  Button,
+  Form,
   FormGenerator,
   FormItemConfig,
   FormModal,
+  message,
   RenderType,
 } from 'art-antd-react';
-import React, { useState } from 'react';
-//  安裝了 art-antd-react 以后 antd 直接导出的 类型或组件 可以直接从 art-antd-react 中导出
-// import { FormItemConfig, RenderType, useFormModal, Button, message } from 'art-antd-react';
+import { useState } from 'react';
 
 const FormModalDemo1 = () => {
   const [visible, setVisible] = useState(false);